Refactored the RTMP codec to decouple from Connection. The state is now stored in an RTMP object attached as an attribute to the IoSession. This work is in advance of further refactoring of the Connection and Client objects coming soon.
Updated the DebugProxyHandler to work with new mina setup. The proxy is now implemented as a filter which is a much more flexible way of doing things. 

Added ChunkSize 0x01 RTMP type. In the protocol decoder this now effects the chunk size for the rest of the IoSession. The value is stored on the RTMP state object. The red5 server does not currently send ChunkSize while streaming. We need to work out if it works with flash 7 or its just a flash 8 thing and also what values may be optimal.

Deleted some old classes that were not doing anything. Some of these will come back as the api is implemented.